mobileness

+++ ctl/media_ctl.class.php
@@ -8,26 +8,40 @@ class media_ctl {
 
     define('KDEBUG_JSON', true);
 
-    if (isset($_REQUEST['signed_request'])) {
+    if (isset($_REQUEST['signed_request']) && !empty($_REQUEST['signed_request'])) {
+
       $fb = new fb();
-      if (isset($fb->session['page']) && isset($fb->session['page']['admin'])) {
+
+      if (
+        isset($fb->session['page']) && 
+        isset($fb->session['page']['admin']) && 
+        isset($fb->session['app_data']) && 
+        $fb->session['page']['admin'] == true &&
+        $fb->session['app_data'] == 'admin'
+        ) {
         $this->admin = true;
       }
-    }
 
+    }
 
   }
 
   public function index($date) {
 
-    $query = 'ORDER BY created DESC LIMIT 20';
+    if ($this->admin == true) {
+      $filter = ' WHERE (flagged = 1 OR flagged = 0)';
+    } else {
+      $filter = ' WHERE flagged = 0';
+    }
+
+    $order = ' ORDER BY created DESC LIMIT 20';
 
     if (isset($_REQUEST['last']) && !empty($_REQUEST['last'])) {
-      $query = 'WHERE created < "'.$_REQUEST['last'].'" '.$query;
+      $filter .= ' AND created < "'.$_REQUEST['last'].'" ';
     }
 
     $data = array();
-    foreach (cache::gets($query) as $cache) {
+    foreach (cache::gets($filter.$order) as $cache) {
       $data[] = $cache->data();
     }
 
@@ -66,6 +80,28 @@ class media_ctl {
     }
 
     $single->flagged = !$single->flagged;
+    $single->save();
+
+    echo json_encode(['error' => false]);
+    return true;
+
+  }
+
+  public function stick($id) {
+
+    if ($this->admin != true) {
+      echo json_encode(['error' => true, 'message' => 'not an admin']);
+      return true;
+    }
+
+    $single = new cache($id);
+
+    if (!$single->exists()) {
+      echo json_encode(['error' => true]);
+      return true;
+    }
+
+    $single->stuck = !$single->stuck;
     $single->save();
 
     echo json_encode(['error' => false]);
